requirements.txt
Django>=5.0,<5.1.0
djangorestframework
Pillow>=10.1.0,<10.2.0